misc: MSVC build fixes
* Warning fix

taking the negative of a unsigned value is still positive on MSVC it seems. So I cast it to a signed value first.

* MSVC weirdness

for some reason maths.h doesn't define M_PI and you have to set this define to make it
Also in the ongoing war with "things other than MS" it will complain that strdup is not a true C standard name and you have to use _strdup, which gcc won't accept, so I just typedef.

* cannot perform maths on type void

gcc defaults void to a 1 byte pointer, so I cast to a char* here to perform the same result on other compilers.

* remove the mess VS makes

This removed all of the VS build folders and debug output. I've also included the sln, at the moment I don't think it is ready for everybody to use.
Fixed paths in the sln ( not sure if I can solve it, but there might be a better way to solve it )
Doesn't run the python build step, not a deal breaker for the most part but somebody doing a fresh grab would end up running into the issue and this will cause  problems for them.
But this will allow me to move between branches and cherry pick check-ins again.

* change GameAllocMemPointer to char*

This allows the + and -  maths done on alloc and free to compile in all compilers.

+++ src/game/lara.c
@@ -2027,7 +2027,7 @@ int32_t TestLaraVault(ITEM_INFO *item, COLL_INFO *coll)
         item->anim_number = AA_STOP;
         item->frame_number = AF_STOP;
         Lara.calc_fall_speed =
-            -(phd_sqrt((int)(-2 * GRAVITY * (hdif + 800))) + 3);
+            -(int16_t)(phd_sqrt((int)(-2 * GRAVITY * (hdif + 800))) + 3);
         AnimateLara(item);
         item->pos.y_rot = angle;
         ShiftItem(item, coll);
